rabbit_mnesia_rename: Use the new rabbit_nodes_common:name_type() function
... instead of calling rabbit_nodes:name_type(). The latter now uses the
context to return the name type of the current node (instead of the
environment variable which may not be set). However, this function is
executed in the context of the CLI which does not start
rabbit_prelaunch. Therefore, there is no context to get the name type
from.

Anyway, the name type should be determined based on the node name we are
about to impersonate. So use the argument to deduce the name type.